requirements.txt
# ===========================================
# LangChain & LangGraph Practice - Requirements
# ===========================================

# Core LangChain
langchain>=0.3.0
langchain-core>=0.3.0
langchain-community>=0.3.0

# LangGraph - Agent Framework
langgraph>=0.2.0
langgraph-checkpoint>=2.0.0
langgraph-checkpoint-sqlite>=2.0.0

# LLM Provider Integrations
langchain-openai>=0.2.0
langchain-anthropic>=0.2.0
langchain-google-genai>=2.0.0
langchain-groq>=0.2.0
langchain-cohere>=0.3.0
langchain-huggingface>=0.1.0

# Direct SDK Dependencies
openai>=1.0.0
google-generativeai>=0.8.0
huggingface-hub>=0.20.0
anthropic>=0.25.0
groq>=0.9.0
cohere>=5.0.0

# Document Processing
pypdf>=3.0.0
python-magic-bin>=0.4.14; sys_platform == 'win32'

# Text Splitting
tiktoken>=0.5.0

# Vector Stores & Embeddings
faiss-cpu>=1.7.4

# Web Search & Tools
duckduckgo-search>=3.0.0
tavily-python>=0.5.0

# HTTP & Web
requests>=2.31.0
aiohttp>=3.9.0
httpx>=0.25.0

# Environment & Config
python-dotenv>=1.0.0

# Web UI
streamlit>=1.25.0

# Jupyter Notebooks
ipykernel>=6.29.0

# Data Processing
numpy>=1.24.0
pandas>=2.0.0

# Async Support
anyio>=4.0.0

# JSON & Data Validation
pydantic>=2.0.0
pydantic-settings>=2.0.0
jsonpatch>=1.33

# SQLite for LangGraph Checkpoints
aiosqlite>=0.19.0

# Logging & Monitoring
langsmith>=0.1.0
tenacity>=8.2.0